techpack/display: sde: Skip heavy autorefresh checks when it's not enabled
These heavy checks for seeing if autorefresh is enabled are unneeded
when the autorefresh config is disabled. These checks are performed on
every display commit and show up as using a significant amount of CPU
time in perf top. Skip them when it's unnecessary in order to improve
display rendering performance.

[ hridaya: Adapt for k4.19 ]

Change-Id: I7ecc169f4671409074092045c02d42f2e6148c79
Signed-off-by: Sultan Alsawaf <sultan@kerneltoast.com>
Signed-off-by: hridaya <hridaya@pixelos.net>
Signed-off-by: NotZeetaa <rodrigo2005contente@gmail.com>